Some types are changes for formatting.

+++ TeaCoin/Commons/Presentations/Extension/CoinPresentation+API.swift
@@ -11,6 +11,6 @@ import TeaCoinAPI
 extension CoinPresentation {
     
     convenience init(coin: CoinProperties) {
-        self.init(symbol: coin.symbol, name: coin.name, price: coin.price, change: coin.change, icon: coin.iconUrl)
+        self.init(symbol: coin.symbol, name: coin.name, price: coin.price.toDouble(), change: coin.change.toDouble(), icon: coin.iconUrl)
     }
 }
